requirements.txt
# requirements.txt

requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.1
flask==2.0.1
werkzeug==2.0.3
schedule==1.2.0